Removed: global filters and instead defined utility methods and updated the usage in components(#1wut90a)
